moved request reading from Server into  RequestReader class.
