+++ public/form.html
@@ -1,0 +1,20 @@
+<!doctype html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport"
+          content="width=device-width, user-scalable=no, initial-scale=1.0, maximum-scale=1.0, minimum-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Document</title>
+</head>
+<body>
+<form action="http://localhost:9999" method="get">
+    <input> <!-- without name - никуда не отправится -->
+    <input name="title">
+    <input name="value">
+    <input name="value"> <!-- специально два поля с одним name -->
+    <input type="file" name="image"> <!-- для отправки файлов -->
+    <button>Send</button>
+</form>
+</body>
+</html>
