+++ src/main/java/Main.java
@@ -5,10 +5,12 @@ import java.nio.file.Path;
 import java.time.LocalDateTime;
 
 /**
- * Server demo class. Registers two handlers and launches listening.
+ * Server demo class. Registers several handlers and launches listening.
  * The first handler is specific full-path handler, it triggers only when client requests exactly
+ * for path specified in handler path ("/" which means http://localhost:9999).
+ * The second handler is specific full-path handler, it triggers only when client requests exactly
  * for path specified in handler path (relatively to PUBLIC_FOLDER).
- * The second handler is "common" which means it is only seeked for if no full-path handlers found
+ * The third handler is "common" which means it is only seeked for if no full-path handlers found
  * for request path, and it handles all request paths whose parent folder is specified in handler path.
  * In this example "common" handler will handle all requests for resources located in PUBLIC_FOLDER
  * except classic.html, which is handled by specific full-path handler
@@ -20,6 +22,29 @@ public class Main {
 
     public static void main(String[] args) {
         final var server = new Server();
+
+        //Custom exact match handler for root path like http://localhost:9999"
+        server.addHandler(Method.GET, "/", (r, o) -> {
+            try {
+                final Path filePath = Path.of(".", PUBLIC_FOLDER, "/index.html");
+                final String mimeType = Files.probeContentType(filePath);
+                final var length = Files.size(filePath);
+                o.write(Server.buildResponseStatusHeadersOnOK(
+                        mimeType, length, !r.headerExists("Connection: keep-alive")).getBytes());
+                Files.copy(filePath, o);
+                o.flush();
+            } catch (InvalidPathException | IOException e) {
+                e.printStackTrace();
+                //Not found. Respond with 404
+                try {
+                    o.write(Server.buildResponseStatusHeadersOnFail(
+                            !r.headerExists("Connection: keep-alive")).getBytes());
+                    o.flush();
+                } catch (IOException ex) {
+                    ex.printStackTrace();
+                }
+            }
+        });
 
         //Custom exact match handler for PUBLIC_FOLDER/classic.html"
         server.addHandler(Method.GET, "/classic.html", (r, o) -> {
