+++ src/main/java/Server.java
@@ -75,7 +75,7 @@ public class Server {
              final var out = new BufferedOutputStream(socket.getOutputStream())) {
 
             // read request string into Request object
-            Request request = readRequest(in);
+            Request request = RequestReader.readRequest(in, REQUEST_LENGTH_LIMIT);
             if (request == null) {
                 // just close socket
                 System.out.println("closing socket");
@@ -84,13 +84,16 @@ public class Server {
             }
 
             //find registered handler exactly matching handlers path
-            Handler handler = handlerMap.get(request.getMethod()).get(request.getPath());
-            System.out.println("request path: " + request.getPath());
+            Handler handler = handlerMap
+                    .get(request.getMethod())
+                    .get(request.getPathWithoutQueryString());
+            System.out.println("request full path: " + request.getPath());
+            System.out.println("request path: " + request.getPathWithoutQueryString());
 
-            //if exact match not found, try to find handler by parent path to request path
-            if (handler == null) {
+            //if exact match not found, and path is not root /, try to find handler by parent path to request path
+            if (handler == null && !request.getPathWithoutQueryString().equals("/")) {
                 System.out.println("handler not found for exact request path");
-                Path parent = Path.of(request.getPath()).getParent();
+                Path parent = Path.of(request.getPathWithoutQueryString()).getParent();
                 System.out.println("parent path: " + parent);
 
                 //if this one fails either - no registered handlers found, respond with 404
@@ -117,86 +120,7 @@ public class Server {
         }
     }
 
-    /**
-     * Reads http request from input stream, validates it and returns it as the Request object
-     * @param in - BufferedInputStream of client socket
-     * @return Request object if reading/parsing was ok, null if not (bad request)
-     */
-    private static Request readRequest (BufferedInputStream in) {
-        try {
-            in.mark(REQUEST_LENGTH_LIMIT);
-            final var buffer = new byte[REQUEST_LENGTH_LIMIT];
-            final var read = in.read(buffer);
-
-            // look for request line end
-            final var requestLineDelimiter = new byte[]{'\r', '\n'};
-            final var requestLineEnd = indexOf(buffer, requestLineDelimiter, 0, read);
-            if (requestLineEnd == -1) {
-                return null;
-            }
-
-            // read request line
-            final var requestLine = new String(Arrays.copyOf(buffer, requestLineEnd)).split(" ");
-            if (requestLine.length != 3) {
-                return null;
-            }
-
-            final Method method;
-            try {
-                method = Enum.valueOf(Method.class,requestLine[0]);
-            } catch (IllegalArgumentException e) {
-                // bad request;
-                return null;
-            }
-            System.out.println(method);
-
-            final var path = requestLine[1];
-            if (!path.startsWith("/")) {
-                // bad request;
-                return null;
-            }
-            System.out.println(path);
-
-            // ищем заголовки
-            final var headersDelimiter = new byte[]{'\r', '\n', '\r', '\n'};
-            final var headersStart = requestLineEnd + requestLineDelimiter.length;
-            final var headersEnd = indexOf(buffer, headersDelimiter, headersStart, read);
-            if (headersEnd == -1) {
-                // bad request;
-                return null;
-            }
-
-            // отматываем на начало буфера
-            in.reset();
-            // пропускаем requestLine
-            in.skip(headersStart);
-
-            final var headersBytes = in.readNBytes(headersEnd - headersStart);
-            final var headers = Arrays.asList(new String(headersBytes).split("\r\n"));
-            System.out.println(headers);
-
-            // определяем тело, для GET тела нет
-            byte[] bodyBytes = new byte[0];
-            if (!method.equals(Method.GET)) {
-                in.skip(headersDelimiter.length);
-                // вычитываем Content-Length, чтобы прочитать body
-                final var contentLength = Request.extractHeader(headers, "Content-Length");
-                if (contentLength.isPresent()) {
-                    final var length = Integer.parseInt(contentLength.get());
-                    bodyBytes = in.readNBytes(length);
-                    final var body = new String(bodyBytes);
-                    System.out.println(body);
-                }
-            }
-
-            return new Request(method, path, headers, bodyBytes);
-        } catch (IOException e) {
-            e.printStackTrace();
-            return null;
-        }
-    }
-
-    /**
+     /**
      * Builds response status line and headers for OK status
      * @param mimeType - mimetype of response body
      * @param length - respince body length
@@ -228,20 +152,6 @@ public class Server {
                 "Content-Length: 0\r\n" +
                 "Connection: close\r\n" +
                 "\r\n";
-    }
-
-    // from google guava with modifications
-    private static int indexOf(byte[] array, byte[] target, int start, int max) {
-        outer:
-        for (int i = start; i < max - target.length + 1; i++) {
-            for (int j = 0; j < target.length; j++) {
-                if (array[i + j] != target[j]) {
-                    continue outer;
-                }
-            }
-            return i;
-        }
-        return -1;
     }
 }
